fixed some functions and created striteri

+++ libft/ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: herbie <herbie@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/07 12:16:23 by hvine             #+#    #+#             */
-/*   Updated: 2022/11/10 11:08:20 by herbie           ###   ########.fr       */
+/*   Updated: 2022/11/11 15:31:54 by herbie           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	if (n <= 0)
